responses & tests

+++ src/main/java/edu/eci/ieti/triddy/services/impl/NotificationServiceImpl.java
@@ -1,10 +1,13 @@
 package edu.eci.ieti.triddy.services.impl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import edu.eci.ieti.triddy.exceptions.NotificationNotFoundException;
+import edu.eci.ieti.triddy.exceptions.UserNotFoundException;
 import edu.eci.ieti.triddy.model.Notification;
 import edu.eci.ieti.triddy.model.User;
 import edu.eci.ieti.triddy.repository.NotificationRepository;
@@ -21,31 +24,45 @@ public class NotificationServiceImpl implements NotificationService{
     private UserRepository userRepository;
 
     @Override
-    public List<Notification> getNotifications(String email) {
-        return notificationRepository.findByUser(email);
+    public List<Notification> getNotifications(String email) throws UserNotFoundException {
+        User user = userRepository.findByEmail(email);
+        if (user != null){
+            return notificationRepository.findByUser(email);
+        }else{
+            throw new UserNotFoundException(email);
+        }
     }
 
     @Override
-    public Notification setNotification(Notification notification){
+    public Notification setNotification(Notification notification) throws UserNotFoundException{
         User user = userRepository.findByEmail(notification.getUser());
         if (user != null){
             return notificationRepository.save(notification);
         }else{
-            return null;
+            throw new UserNotFoundException(notification.getUser());
+        }
+    }
+
+    @Override
+    public void delNotification(String notifId) throws NotificationNotFoundException {
+        Optional<Notification> notification = notificationRepository.findById(notifId);
+        if (notification.isPresent()){
+            notificationRepository.deleteById(notifId);
+        }else{
+            throw new NotificationNotFoundException(notifId);
         }
         
     }
 
     @Override
-    public void delNotifications(List<String> notifIds) {
-        for (String str: notifIds){
-            notificationRepository.deleteById(str);
+    public void delNotificationsUser(String user) throws UserNotFoundException{
+        User user2 = userRepository.findByEmail(user);
+        if (user2 != null){
+            notificationRepository.deleteByUser(user);
+        }else{
+            throw new UserNotFoundException(user);
         }
-    }
-
-    @Override
-    public void delNotificationsUser(String user){
-        notificationRepository.deleteByUser(user);
+        
     }
     
 }
